fix: moved logic into createHelpHeaderButton
